fix TODO move WaitForPodCondition to framework/pods
